Major change was adding a contextual action mode for deleting items. Also fixed bug with enforcing a unique name.
FIX: NameModifier makeNameUnique was incorrectly enforcing uniqueness. Used a temporary string to hold the potential unique name. Previously, whenever a digit was added to the name, it used the name before it, which might have a digit already. For example John 1 might become John 1 1.
ADD: Added contextual action mode for deleting items from a group item list.
ADD: Instead of an edit button in the action bar, an item is clicked to edit it.