Added contextual action mode to group list
The group list uses action bar for edit and delete actions. Both are
moved to the contextual action mode. Using the contextual action
mode does the same thing as the action bar implementation, but uses
a lot less code. Help toast messages were also added to both the
group list and the group item list to help users be aware of the
contextual action mode/long-click.