Added HitDie class and test units for die evaluator
The HitDie class evaluates die expressions represented as a string.
This is used for the creature catalog, where creatures can represent
max hit points as either a hit dice expression or integer. A test
class was also added to test the HitDie class. In the creature catalog
database, the MaxHP column data type was changed to VARCHAR since it can
now represent either a hit die expression or integer.