Changed Creature member types from int to string
Most of the Creature classes private members such as ability scores was
changed from integer to String. Although integers more naturally
represent ability scores and most of a creature's statistics, the data
is mostly used for displaying stat information and very little
calculation is done. There was much parsing and conversion of these
members from int to String just to set the text of a TextView or
EditText.

Accompanying this change was a change as well to the database collumn
type from INTEGER to TEXT.

Parse and conversion functions such as String.valueOf(...) were removed
since they are no longer needed.