Fixed bug in SettingsActvity due to default preference value
When the options activity is accessed for the very first time,
onSharedPreferenceChanged method is called if a default value is set for a
preference. The view is not created yet, so a null progress bar was being passed
onto the LoadCatalogTask constructor resulting in a crash. This problem
was fixed by removing the default value so that onSharedPreferenceChanged is not
called before the view for the preferences are created, and also a check
for a null progress bar was made.