Implemented Creature Catalog Search
User can search for entries in the Creature Catalog. If the query is
1 to 3 letters long, it only matches if an item in the catalog starts
with the query. If query length is > 3, then a catalog item matches
if it starts with OR contains the query. The matches are then
displayed. If the search is collapsed, the selected catalog is reloaded
including any changes that may have been made while a filtered list
was shown.