Validation of hit die expression in creature catalog
A valid hit die expression was made in the creature catalog when adding
or deleting a creature. If a hit die is invalid, an toast error message
is shown. Toast messages common to multiple activities are collected in
CommonToast class as static methods.

A minor change to GroupListActivity was made: toast message when
cancelling a new or rename of party/encounter was previously
"party name required" and is now "<groupType> name required" where
<groupType> is either "party" or "encounter".