Creature from scratch can be added to initiative
A creature can now be added to initiative by clicking the proper action
bar button.

The dialog used to edit creatures was heavily modified to add support
for use with every activity, including the initiative activity. Certain
views are hidden or exposed depending on the activity the dialog is
being used in.