Changed layout of how creature stats are displayed
groupitemdisplay.xml used a relative layout to place the initiative and
hit point views. Both are now joined with the linearlayout holding the
armor class stat. The reason for the change is that a relative layout
did not look good when the phone is turned to a horizontal view. The
relative layout caused the init and hp views to be misaligned with the
other views.