Database changes due to new hit die property
A new hit die property was added to the Creature class. The reason is
that a distinguishment between max hit points and hit die needed to be
made for properly displaying current/max hit points in the
InitiativeActivity. This resulted in database changes: adding hit die
info to any table that held creature information.

As a result of database changes, adapters and dialogs that work with the
adapters needed to be changed to use properly display either hit die or
max hit points depending on the context.