Add parties to initiative and maximize hit die roll
Parties can now be added to the initiative list by selecting
Add->Party and selecting a party from the parties list.

For parties, if a party item has a hit die expression, then the max hit
points is set by maximizing the hit die value. A new method getMaxVal()
is available in the HitDie class to get the maximum value of a die
expression.