Added an input filter for hit point EditText
An input filter was added for the max hit point EditText that limits
characters to only those valid for a hit die expression. Previously, a
key listener was being used, and although this worked for hardware
keyboards, it did not work for soft/virtual keyboards.

A note about inputType in groupitemedit.xml:
inputType "textnoSuggestions" does not work for all devices. Although a
bit of a hack, the additional "textVisiblePassword" in another layer
that prevents word suggestions when typing into an edit text. Testing
was done on a samsung galaxy tab 2 7'' device and "textNoSuggestions"
was not working, but with "textVisiblePassword" it works.