Coloring of PCs/Monsters in initiative
PCs and monster creatures are now distinguished in the initiative list by
color. Names for PCs are green and red for monsters.

A related bug fix was also made. The CreatureEditDialog now uses
isChecked() for the creatureTypeSwitch Switch widget. Previously,
getText() was being used, but the method always returns an empty string.
I believe this is a bug in the Android API.