Bug fix: PC name set to monster color in initiative
PC names should be green and monsters red. In the initiative activity,
when a monster(s) was sandwiched in between two PCs, if the monster was
removed from the list, the PC that took its place would now have a red
colored text for its name, which is incorrect behavior.

The issue was due to the adapter reusing a view for a certain position.
If the view is reused, it holds its old values, thus a new value needs to
be set. An else statement was added to account for the case in which the
view gets reused, but needs a different value, in this case, text color.

Also, a creature object was changed to not be a monster by default, so
that its name text shows up as green instead of red in the pc and
encounter manager.