Delete creatures in initiative and update bug fix
Creatures in initiative can be deleted selectively by long clicking an
item, optionally selecting more items, and then clicking on the discard
button.

A bug fix was also made: updating a creature's hit points was not being
saved into the database correctly. The issue was that currentHitPoints
was not being saved as a value into the database.